ref(normalization): Restore span processing to transactionprocessor
